scripts/kafka_producer/requirements.txt
kafka-python
pandas
python-dotenv
